Revert "Ho fatto delle modifiche al tuo codice. Ho rimosso il driver e le proprieta per collegarsi a mysql e ho aggiunto H2 DB per fare dei test."
This reverts commit 8e8367cd8eb0b969a498e912fdb96f2dd60cf93c.

+++ filmlist/src/main/java/com/example/filmlist/repository/FilmRepository.java
@@ -3,9 +3,9 @@ package com.example.filmlist.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.filmlist.entity.Film;
+import com.exemple.filmlist.models.Film;
 
 @Repository
-public interface FilmRepository extends CrudRepository<Film, Long> {
+public interface FilmRepository extends CrudRepository<Film, Integer> {
 
 }
